add autoindent feature
this is mainly intended for lisp code, but could be useful in other
files also I guess. the idea is that when you open a line (usually using
o in normal mode) immediately below the end of an expression, you do
actually want the line indented to match the previous line, so you can
keep adding stuff to that expression (parinfer fixes the parens as you
soon as you start typing). if you on the other hand actually want to
start writing on a new expression, it makes sense to me that you
wouldn't really want that so close to the previous expression, you'd
rather have a blank line between them (in which case this plugin doesn't
activate and everything behaves as normal).